+++ scripts/shipyard_nodeprep.sh
@@ -598,7 +598,7 @@ EOF
         -v $AZ_BATCH_NODE_ROOT_DIR:$AZ_BATCH_NODE_ROOT_DIR \
         -w $AZ_BATCH_TASK_WORKING_DIR \
         -p 6881-6891:6881-6891 -p 6881-6891:6881-6891/udp \
-        alfpark/batch-shipyard:cascade-latest
+        alfpark/batch-shipyard:cascade-latest &
     cascadepid=$!
 else
     # backfill node prep start